Organize it a bit more
