+++ src/components/splashImage/SplashImage.jsx
