+++ src/components/splashImage/SplashImage.module.css
