+++ src/views/Homepage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "./Homepage.module.css";
 
-import SplashImage from "../components/SplashImage";
+import SplashImage from "../components/splashImage/SplashImage";
 
 const Homepage = () => {
   return (
